Train value network + misc. fixes
